cdrom-handler: Fix some debug prints on x86_64

git-svn-id: https://svn.aros.org/svn/aros/trunk@43557 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/filesys/CDVDFS/src/cdrom.c
@@ -276,11 +276,11 @@ int Read_From_Drive
     p_cd->scsireq->io_Actual   = (ULONG) p_sector >> 21;
     p_cd->scsireq->io_Command  = p_cd->scsireq->io_Actual ? TD_READ64 : CMD_READ;
 
-    D(bug("[CDVDFS]\tAccessing sectors %ld:%ld\n", p_sector, p_number_of_sectors));
+    D(bug("[CDVDFS]\tAccessing sectors %ld:%ld\n", (long)p_sector, (long)p_number_of_sectors));
 
     if (0 != DoIO((struct IORequest*) p_cd->scsireq))
     {
-	D(bug("[CDVDFS]\tTransfer failed: %ld\n", p_cd->scsireq->io_Error));
+	D(bug("[CDVDFS]\tTransfer failed: %ld\n", (long)p_cd->scsireq->io_Error));
         Clear_Sector_Buffers(p_cd);
 	return 0;
     }
